JackSocket: Continue read/write if only partly done
If a message was only written partly this has not to be interpreted as a
fatal error.
With this change it tries to continue reading/writing as long as there
are bytes read/written.

In addition errno is only valid in case of a return value of -1.
Therefore errno will only be read on a negative error value, now.

Change-Id: I4b46142c480e016b187bc606b248176eb23aefd8
Signed-off-by: Timo Wischer <twischer@de.adit-jv.com>